pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "markdown-pdf"
version = "0.1.0"
description = "Conversion Markdown → PDF via LaTeX avec personnalisation"
readme = "README.md"
authors = [{name = "Redwan Kara"}]
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "Jinja2>=3.1",
    "typer>=0.12",
    "pydantic>=2.5",
    "PyYAML>=6.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "ruff>=0.3",
    "black>=24.0"
]

[project.scripts]
markdown-pdf = "markdown_pdf.cli:app"

[tool.setuptools]
package-dir = {"" = "src"}
include-package-data = true

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
markdown_pdf = [
    "templates/*.tex.j2",
]
